Implementing user roles

+++ app.py
@@ -2,10 +2,11 @@ from flask import Flask, request, jsonify
 from models.user import User
 from flask_login import LoginManager, login_user, current_user, logout_user, login_required
 from database import db
+import bcrypt
 
 app = Flask(__name__)
 app.config['SECRET_KEY'] = 'your_secret_key'
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///database.db'
+app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql+pymysql://root:admin123@127.0.0.1:3306/flask-crud'
 
 login_manager = LoginManager()
 db.init_app(app)
@@ -27,7 +28,7 @@ def login():
         #login
         user = User.query.filter_by(username=username).first()
 
-        if user and user.password == password:
+        if user and bcrypt.checkpw(str.encode(password), str.encode(user.password)):
             login_user(user)
             print(current_user.is_authenticated)
             return jsonify({"message": "Authentication successfull"})
@@ -47,7 +48,8 @@ def create_user():
     password = data.get("password")
 
     if username and password:
-        user = User(username=username, password=password)
+        hashed_password = bcrypt.hashpw(str.encode(password), bcrypt.gensalt())
+        user = User(username=username, password=hashed_password, role='user')
         db.session.add(user)
         db.session.commit()
         return jsonify({"message": f"{username} successfully registered"})
@@ -70,6 +72,9 @@ def update_user(id_user):
     data = request.json
     user = User.query.get(id_user)
 
+    if id_user != current_user.id and current_user.role == 'user':
+        return jsonify({"message":"You do not have permission to perform this action."}), 403
+
     if user and data.get("password"):
         user.password = data.get("password")
         db.session.commit()
@@ -81,6 +86,9 @@ def update_user(id_user):
 @login_required
 def delete_user(id_user):
     user = User.query.get(id_user)
+
+    if current_user.role != 'admin':
+        return jsonify({"message":"You do not have permission to perform this action."}), 403
 
     if user and id_user != current_user.id:
         db.session.delete(user)
